+++ R/tostring.R
@@ -1,43 +1,52 @@
-## this can't be tested from within R
-# nocov start
-#' @importFrom stats na.omit
-#' @importFrom utils head tail localeToCharset
-#' @import checkmate
-
-d_hsep_factory <- function() {
-  warn_sent <- FALSE
-  function() {
-    if (any(grepl("^UTF", localeToCharset()))) {
-      "\u2014"
-    } else {
-      if (!warn_sent && interactive()) {
-        message(
-          "Detected non-UTF charset. Falling back to '-' ",
-          "as default header/body separator. This warning ",
-          "will only be shown once per R session."
-        )
-        warn_sent <<- TRUE
-      }
-      "-"
-    }
-  }
-}
-
-#' Default horizontal Separator
+#' @title Default horizontal separator
 #'
-#' The default horizontal separator character which can be
+#' @description The default horizontal separator character which can be
 #' displayed in the current `charset` for use in rendering table-likes.
+#'
+#' @param hsep_char character(1). Character that will be set in the R environment
+#'   options as default for creating the horizontal separator. It needs to be
+#'   single character. Use `getOption("formatters_default_hsep")` to get its current
+#'   value (`NULL` if not set).
 #'
 #' @return `unicode` 2014 (long dash for generating solid horizontal line)
 #' if in a locale that uses a UTF character set, otherwise an ASCII hyphen
 #' with a once-per-session warning.
 #'
-#' @export
 #' @examples
 #' default_hsep()
-default_hsep <- d_hsep_factory()
+#' set_default_hsep("o")
+#' default_hsep()
+#'
+#' @name default_horizontal_sep
+#' @export
+default_hsep <- function() {
+  system_default_hsep <- getOption("formatters_default_hsep")
 
-# nocov end
+  if (is.null(system_default_hsep)) {
+    if (any(grepl("^UTF", utils::localeToCharset()))) {
+      hsep <- "\u2014"
+    } else {
+      if (interactive()) {
+        warning(
+          "Detected non-UTF charset. Falling back to '-' ",
+          "as default header/body separator. This warning ",
+          "will only be shown once per R session."
+        ) # nocov
+      } # nocov
+      hsep <- "-" # nocov
+    }
+  } else {
+    hsep <- system_default_hsep
+  }
+  hsep
+}
+
+#' @name default_horizontal_sep
+#' @export
+set_default_hsep <- function(hsep_char) {
+  checkmate::assert_character(hsep_char, n.chars = 1, len = 1, null.ok = TRUE)
+  options("formatters_default_hsep" = hsep_char)
+}
 
 .calc_cell_widths <- function(mat, colwidths, col_gap) {
   spans <- mat$spans
@@ -383,17 +392,19 @@ decimal_align <- function(string_mat, align_mat) {
 #' `max_width`), or horizontal separator character (e.g. `hsep = "+"`).
 #'
 #' @inheritParams MatrixPrintForm
-#' @param widths numeric (or  NULL). (proposed) widths for the columns
+#' @param widths numeric (or  `NULL`). (proposed) widths for the columns
 #'     of \code{x}. The expected length  of this numeric vector can be
 #'     retrieved with  `ncol() + 1`  as the  column of row  names must
 #'     also be considered.
 #' @param hsep character(1). Characters to repeat to create
-#'     header/body separator line.
+#'     header/body separator line. If `NULL`, the object value will be
+#'     used. If `" "`, an empty separator will be printed. Check [default_hsep()]
+#'     for more information.
 #' @param tf_wrap logical(1). Should  the texts for  title, subtitle,
 #'     and footnotes be wrapped?
-#' @param max_width integer(1), character(1) or NULL. Width that title
+#' @param max_width integer(1), character(1) or `NULL`. Width that title
 #'     and   footer   (including   footnotes)  materials   should   be
-#'     word-wrapped to. If NULL, it is  set to the current print width
+#'     word-wrapped to. If `NULL`, it is  set to the current print width
 #'     of the  session (`getOption("width")`). If set to `"auto"`,
 #'     the width of the table (plus any table inset) is used. Ignored
 #'     completely if `tf_wrap` is `FALSE`.
@@ -422,7 +433,7 @@ setMethod("toString", "MatrixPrintForm", function(x,
                                                   tf_wrap = FALSE,
                                                   max_width = NULL,
                                                   col_gap = mf_colgap(x),
-                                                  hsep = default_hsep()) {
+                                                  hsep = NULL) {
   checkmate::assert_flag(tf_wrap)
 
   mat <- matrix_form(x, indent_rownames = TRUE)
@@ -517,6 +528,9 @@ setMethod("toString", "MatrixPrintForm", function(x,
 
   # Define gap string and divisor string
   gap_str <- strrep(" ", col_gap)
+  if (is.null(hsep)) {
+    hsep <- horizontal_sep(mat)
+  }
   div <- substr(strrep(hsep, ncchar), 1, ncchar)
   hsd <- header_section_div(mat)
   if (!is.na(hsd)) {